+++ internal/app/wwctl/configure/tls/main_test.go
@@ -1,4 +1,4 @@
-package keys
+package tls
 
 import (
 	"bytes"
